+++ kdroid/.gitignore
