+++ kdroid/build.gradle
