+++ kdroid/proguard-rules.pro
