+++ kdroid/src/main/AndroidManifest.xml
