+++ kdroid/src/main/java/org/chx/kdroid/candy/context/ActivityCandy.kt
@@ -1,0 +1,11 @@
+package org.chx.kdroid.candy.context
+
+import android.app.Activity
+import android.content.Context
+import android.content.Intent
+import kotlin.reflect.KClass
+
+
+fun <A : Activity> Context.startActivity(cls: KClass<A>, func: Intent.() -> Unit = {}) {
+    startActivity(Intent(this, cls.java).apply { func() })
+}
