+++ kdroid/src/main/java/org/chx/kdroid/jadapter/ViewHolder.kt
