+++ kdroid/src/main/java/org/chx/kdroid/jadapter/adapter/YaRecyclerAdapter.kt
@@ -1,0 +1,24 @@
+package org.chx.kdroid.jadapter.adapter
+
+import android.support.v7.widget.RecyclerView
+import android.view.ViewGroup
+import org.chx.kdroid.jadapter.ViewHolder
+import org.chx.kdroid.jadapter.proxy.AdapterProxy
+
+class YaRecyclerAdapter<D>(val proxy: AdapterProxy<D>) : RecyclerView.Adapter<YaRecyclerAdapter.YaViewHolder<D>>() {
+
+    override fun getItemCount() = proxy.size
+
+    override fun getItemViewType(position: Int) = proxy.getViewType(position)
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): YaViewHolder<D> {
+        val viewHolder = proxy.getViewHolderWithViewType(parent, viewType)
+        return YaViewHolder(viewHolder)
+    }
+
+    override fun onBindViewHolder(holder: YaViewHolder<D>, position: Int) {
+        holder.holder.convert(proxy[position])
+    }
+
+    class YaViewHolder<in D>(val holder: ViewHolder<D>) : RecyclerView.ViewHolder(holder.itemView)
+}
