+++ kdroid/src/main/res/values/strings.xml
