+++ sample/build.gradle
@@ -24,7 +24,9 @@ dependencies {
     androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
+    
+    compile project(':kdroid')
+
     compile 'com.android.support:appcompat-v7:25.1.1'
-    compile project(':irefresh')
     testCompile 'junit:junit:4.12'
 }
